Refactor: move contribution models to contributions/models
