+++ app/src/main/java/fr/free/nrw/commons/contributions/ContributionsContract.java
@@ -2,7 +2,7 @@ package fr.free.nrw.commons.contributions;
 
 import android.content.Context;
 import fr.free.nrw.commons.BasePresenter;
-import fr.free.nrw.commons.data.models.Contribution;
+import fr.free.nrw.commons.contributions.models.Contribution;
 
 /**
  * The contract for Contributions View & Presenter
