+++ app/src/main/java/fr/free/nrw/commons/contributions/ContributionsListContract.java
@@ -1,7 +1,7 @@
 package fr.free.nrw.commons.contributions;
 
 import fr.free.nrw.commons.BasePresenter;
-import fr.free.nrw.commons.data.models.Contribution;
+import fr.free.nrw.commons.contributions.models.Contribution;
 
 /**
  * The contract for Contributions list View & Presenter
