+++ app/src/main/java/fr/free/nrw/commons/repository/UploadRepository.java
@@ -3,7 +3,7 @@ package fr.free.nrw.commons.repository;
 import androidx.annotation.Nullable;
 import fr.free.nrw.commons.category.CategoriesModel;
 import fr.free.nrw.commons.category.CategoryItem;
-import fr.free.nrw.commons.data.models.Contribution;
+import fr.free.nrw.commons.contributions.models.Contribution;
 import fr.free.nrw.commons.contributions.ContributionDao;
 import fr.free.nrw.commons.filepicker.UploadableFile;
 import fr.free.nrw.commons.location.LatLng;
