+++ app/src/main/java/fr/free/nrw/commons/upload/PageContentsCreator.java
@@ -3,7 +3,7 @@ package fr.free.nrw.commons.upload;
 import android.content.Context;
 import androidx.annotation.NonNull;
 import fr.free.nrw.commons.Media;
-import fr.free.nrw.commons.data.models.Contribution;
+import fr.free.nrw.commons.contributions.models.Contribution;
 import fr.free.nrw.commons.filepicker.UploadableFile.DateTimeWithSource;
 import fr.free.nrw.commons.settings.Prefs.Licenses;
 import fr.free.nrw.commons.utils.ConfigUtils;
