+++ app/src/main/java/fr/free/nrw/commons/upload/UploadPresenter.java
@@ -3,7 +3,7 @@ package fr.free.nrw.commons.upload;
 import android.annotation.SuppressLint;
 import fr.free.nrw.commons.CommonsApplication;
 import fr.free.nrw.commons.R;
-import fr.free.nrw.commons.data.models.Contribution;
+import fr.free.nrw.commons.contributions.models.Contribution;
 import fr.free.nrw.commons.filepicker.UploadableFile;
 import fr.free.nrw.commons.kvstore.JsonKvStore;
 import fr.free.nrw.commons.repository.UploadRepository;
